requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
python-multipart==0.0.18
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
recipe-scrapers==15.11.0
fpdf2==2.8.3
requests
anthropic
python-docx
PyMuPDF
Pillow
easyocr
